requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
pyspark
wordcloud
snowflake-connector-python